requirements.txt
python-dotenv
python-telegram-bot
requests
